Modificaciones en el modelo. Crear nuevamente los controladores y las vistas.

+++ SGM/Models/Funcion.cs
@@ -23,7 +23,7 @@ namespace SGM.Models
         public int funcion_id { get; set; }
         public string nombre { get; set; }
         public string descripcion { get; set; }
-        public int EspecialidadId { get; set; }
+        public byte EspecialidadId { get; set; }
     
         public virtual Especialidad Especialidad { get; set; }
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
